app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
edge-tts==6.1.9
gTTS==2.5.4
ffmpeg-python==0.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.109.0
aiofiles==23.2.1
aiohttp==3.9.1
supabase==2.0.3